backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""

readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "fastapi (>=0.119.0,<0.120.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "firebase (>=4.0.1,<5.0.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "docling (>=2.55.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "chromadb (>=1.3.0,<2.0.0)",
    "google-genai (>=1.48.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "ruff (>=0.14.0,<0.15.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]